Refactor OPF reading logic into a separate static class
